Move isomorphism to its own module
